Move escraper/testing to tests/testing
